+++ src/reference.ts
@@ -7,8 +7,8 @@
 /// <reference path="interaction.ts" />
 /// <reference path="label.ts" />
 /// <reference path="renderer.ts" />
-/// <reference path="categoryRenderer.ts" />
 /// <reference path="xyRenderer.ts" />
+/// <reference path="numericXYRenderer.ts" />
 /// <reference path="circleRenderer.ts" />
 /// <reference path="lineRenderer.ts" />
 /// <reference path="barRenderer.ts" />
@@ -17,5 +17,6 @@
 /// <reference path="coordinator.ts" />
 /// <reference path="legend.ts" />
 /// <reference path="chart.ts" />
+/// <reference path="categoryXYRenderer.ts" />
 /// <reference path="categoryBarRenderer.ts" />
 //grunt-end
